Python/Django/MxOnline/requirement.txt
django==1.9
pillow
django-crispy-forms
django-reversion
django-formtools
future
httplib2==0.9.2
six==1.10.0
django-simple-captcha==0.4.6
django-pure-pagination